+++ x/gov/keeper/grpc_query.go
@@ -9,6 +9,8 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
+	kiratypes "github.com/KiraCore/sekai/types"
+	kiraquery "github.com/KiraCore/sekai/types/query"
 	"github.com/KiraCore/sekai/x/gov/types"
 	customstakingtypes "github.com/KiraCore/sekai/x/staking/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
@@ -136,27 +138,38 @@ func (q Querier) Proposals(ctx context.Context, request *types.QueryProposalsReq
 	store := c.KVStore(q.keeper.storeKey)
 
 	var proposals []types.Proposal
+	var pageRes *query.PageResponse
+	var err error
 
 	proposalsStore := prefix.NewStore(store, ProposalsPrefix)
 
-	pageRes, err := query.FilteredPaginate(proposalsStore, request.Pagination, func(key []byte, value []byte, accumulate bool) (bool, error) {
+	onResult := func(key []byte, value []byte, accumulate bool) (bool, error) {
 		var proposal types.Proposal
 		err := q.keeper.cdc.UnmarshalBinaryBare(value, &proposal)
 		if err != nil {
 			return false, err
 		}
-		if request.All || accumulate {
+		if accumulate {
 			proposals = append(proposals, proposal)
 		}
-		return !request.All, nil
-	})
+		return true, nil
+	}
 
-	if err != nil {
-		return nil, sdkerrors.Wrap(types.ErrGettingProposals, fmt.Sprintf("error getting proposals: %s", err.Error()))
+	// we set maximum limit for safety of iteration
+	if request.Pagination != nil && request.Pagination.Limit > kiratypes.PageIterationLimit {
+		request.Pagination.Limit = kiratypes.PageIterationLimit
 	}
 
 	if request.All {
-		pageRes = nil
+		pageRes, err = kiraquery.IterateAll(proposalsStore, request.Pagination, onResult)
+	} else if request.Reverse {
+		pageRes, err = kiraquery.FilteredReversePaginate(proposalsStore, request.Pagination, onResult)
+	} else {
+		pageRes, err = query.FilteredPaginate(proposalsStore, request.Pagination, onResult)
+	}
+
+	if err != nil {
+		return nil, sdkerrors.Wrap(types.ErrGettingProposals, fmt.Sprintf("error getting proposals: %s", err.Error()))
 	}
 
 	res := types.QueryProposalsResponse{